% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_univariate_cutpoints}
\alias{apply_univariate_cutpoints}
\title{Categorize a dataset from the cutpoints returned from Univariate Methods.}
\usage{
apply_univariate_cutpoints(db, quantile_cutpoints, cutpoints)
}
\arguments{
\item{db:}{dataset (data.frame)}

\item{quantile_cutpoints:}{vector of quantile cutpoints (numeric)}

\item{cutpoints:}{vector of cutpoints (numeric)}
}
\value{
list containing discretized dataset (list)
}
\description{
Categorize a dataset from the cutpoints returned from Univariate Methods.
}
