% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_model_selection.R
\name{backward_model_selection}
\alias{backward_model_selection}
\title{Backward Model Selection For Binary Logistic Regression}
\usage{
backward_model_selection(train, test, k_features = 2)
}
\arguments{
\item{train}{Train DataFrame}

\item{test}{Test DataFrame}

\item{k_features}{Stop algorithm at minimum number of features.}
}
\value{
Dataframe where the models are ranked by gini coefficient.
}
\description{
Backward Model Selection For Binary Logistic Regression
}

