% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutpoint_discretization}
\alias{cutpoint_discretization}
\title{Discretize a database with given cutpoints. To be used with cutpoints returned from multivariate disc.}
\usage{
cutpoint_discretization(db, cutpoints)
}
\arguments{
\item{db:}{database (data.frame)}

\item{cutpoints:}{vector of cutpoints (numeric)}
}
\value{
list containing discretized database (list)
}
\description{
Discretize a database with given cutpoints. To be used with cutpoints returned from multivariate disc.
}

