% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilediscretization.R
\name{DiscByQuantile}
\alias{DiscByQuantile}
\title{Quantile Discretization}
\usage{
DiscByQuantile(db, n = 30)
}
\arguments{
\item{db}{Data.}

\item{n}{Number of observations for each quantile.}
}
\description{
Quantile Discretization
}

