% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretize database via topdown greedy approach. Based on discretization package created by HyunJi Kim.}
\usage{
discretize(db, meth = 4, alpha = 0.05, n = 30)
}
\arguments{
\item{db}{database to discretize (data.frame)}

\item{meth}{measures of associations available 1. Caim 2. Cacc 3. Ameva 4. Info. Statistics 5. Kendall's Tau-C
(numeric)}

\item{alpha}{Penalization to limit number of categories. Only for methods 4 and 5. (numeric)}

\item{n}{Desired Number of observations per Quantile.}
}
\value{
list with discretized data, vector of quantile cutpoints and the supervised method cutpoints.
}
\description{
Discretize database via topdown greedy approach. Based on discretization package created by HyunJi Kim.
}
