% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_multivariate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Fit all subsets model on train, evaluate the best model on validation db, then calculates the gini coef
on the test dataframe.}
\usage{
evaluate_model(train, validation, test)
}
\arguments{
\item{train}{Train data (Dataframe)}

\item{validation}{Validation data (Dataframe)}

\item{test}{Test data (Dataframe)}
}
\value{
Gini Coef. of test dataset.
}
\description{
Fit all subsets model on train, evaluate the best model on validation db, then calculates the gini coef
on the test dataframe.
}

