% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_discretization.R
\name{logistic_reg_giniCoef}
\alias{logistic_reg_giniCoef}
\title{Fit logistic regression in the train dataframe and calculates the gini coef. in the test dataframe.}
\usage{
logistic_reg_giniCoef(train, test, respName)
}
\arguments{
\item{train}{Train DataFrame}

\item{test}{Test DataFrame}

\item{respName}{Name of response variable. Necessary to the formula of glm object.}
}
\value{
Returns the gini coef. of the test dataframe.
}
\description{
Fit logistic regression in the train dataframe and calculates the gini coef. in the test dataframe.
}

