% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_multivariate_model.R
\name{modelSelection}
\alias{modelSelection}
\title{We select the best model via Gini Coefficient on validation.}
\usage{
modelSelection(train, validation)
}
\arguments{
\item{train:}{Train Dataframe (DataFrame)}

\item{validation:}{Datafram for validation (DataFrame)}
}
\value{
table containing the ranked models (DataFrame)
}
\description{
We select the best model via Gini Coefficient on validation.
}

