% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_discretization.R
\name{multdiscretization}
\alias{multdiscretization}
\title{Multivariate Discretization.}
\usage{
multdiscretization(train, test)
}
\arguments{
\item{train}{Train Dataframe.}

\item{test}{Test Dataframe.}
}
\value{
Discretized Dataframe.
}
\description{
Multivariate Discretization.
}

