% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_discretization.R
\name{quantile_discretization}
\alias{quantile_discretization}
\title{Discretize a continuous database by quantile.}
\usage{
quantile_discretization(db, n = 30)
}
\arguments{
\item{db}{Dataframe to discretize.}

\item{n}{Numer of observations by quantile.}
}
\value{
List with the discretized dataframe and a vector of cutpoints.
}
\description{
Discretize a continuous database by quantile.
}

