% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_test_split}
\alias{train_test_split}
\title{Separate train and test database.}
\usage{
train_test_split(db, percentual, seed = 999)
}
\arguments{
\item{db}{Dabatase to separate between trainning and test sets (data.frame)}

\item{percentual}{percentual dedicated to test set (numeric)}

\item{seed}{A seed (numeric)}
}
\value{
list containing the databases (list)
}
\description{
Separate train and test database.
}
