% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ameva.R
\name{ameva}
\alias{ameva}
\title{Calculates the Ameva Coefficient}
\usage{
ameva(tb)
}
\arguments{
\item{tb}{Discretization Scheme (table)}
}
\value{
Ameva Coeficient (numeric)
}
\description{
Calculates the Ameva Coefficient
}
